Update vector store examples and add some bug fixes
