+++ samples/semantickernel-concepts/semantickernel-syntax-examples/src/main/java/com/microsoft/semantickernel/samples/syntaxexamples/memory/InMemoryVolatileVectorStore.java
@@ -5,32 +5,24 @@ import com.azure.ai.openai.OpenAIAsyncClient;
 import com.azure.ai.openai.OpenAIClientBuilder;
 import com.azure.core.credential.AzureKeyCredential;
 import com.azure.core.credential.KeyCredential;
-import com.azure.core.util.ClientOptions;
-import com.azure.core.util.MetricsOptions;
-import com.azure.core.util.TracingOptions;
-import com.azure.search.documents.indexes.SearchIndexAsyncClient;
-import com.azure.search.documents.indexes.SearchIndexClientBuilder;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.microsoft.semantickernel.aiservices.openai.textembedding.OpenAITextEmbeddingGenerationService;
-import com.microsoft.semantickernel.connectors.data.azureaisearch.AzureAISearchVectorStoreOptions;
-import com.microsoft.semantickernel.connectors.data.azureaisearch.AzureAISearchVectorStoreRecordCollection;
-import com.microsoft.semantickernel.connectors.data.azureaisearch.AzureAISearchVectorStoreRecordCollectionOptions;
 import com.microsoft.semantickernel.data.vectorsearch.VectorSearchResult;
+import com.microsoft.semantickernel.data.vectorstorage.VectorStoreRecordCollection;
+import com.microsoft.semantickernel.data.VolatileVectorStore;
+import com.microsoft.semantickernel.data.VolatileVectorStoreRecordCollectionOptions;
 import com.microsoft.semantickernel.data.vectorstorage.attributes.VectorStoreRecordDataAttribute;
 import com.microsoft.semantickernel.data.vectorstorage.attributes.VectorStoreRecordKeyAttribute;
 import com.microsoft.semantickernel.data.vectorstorage.attributes.VectorStoreRecordVectorAttribute;
-import java.nio.charset.StandardCharsets;
 import java.util.Arrays;
-import java.util.Base64;
 import java.util.Collections;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
-
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 
-public class AzureAISearchVectorStore {
+public class InMemoryVolatileVectorStore {
 
     private static final String CLIENT_KEY = System.getenv("CLIENT_KEY");
     private static final String AZURE_CLIENT_KEY = System.getenv("AZURE_CLIENT_KEY");
@@ -38,11 +30,7 @@ public class AzureAISearchVectorStore {
     // Only required if AZURE_CLIENT_KEY is set
     private static final String CLIENT_ENDPOINT = System.getenv("CLIENT_ENDPOINT");
 
-    //////////////////////////////////////////////////////////////
-    // Azure AI Search configuration
-    //////////////////////////////////////////////////////////////
-    private static final String AZURE_AI_SEARCH_ENDPOINT = System.getenv("AZURE_AISEARCH_ENDPOINT");
-    private static final String AZURE_AISEARCH_KEY = System.getenv("AZURE_AISEARCH_KEY");
+    // Embedding model configuration
     private static final String MODEL_ID = System.getenv()
         .getOrDefault("EMBEDDING_MODEL_ID", "text-embedding-3-large");
     private static final int EMBEDDING_DIMENSIONS = 1536;
@@ -55,34 +43,42 @@ public class AzureAISearchVectorStore {
         private final String description;
         @VectorStoreRecordDataAttribute
         private final String link;
-        @VectorStoreRecordVectorAttribute(dimensions = EMBEDDING_DIMENSIONS, indexKind = "Hnsw")
+        @VectorStoreRecordVectorAttribute(dimensions = EMBEDDING_DIMENSIONS, indexKind = "Hnsw", distanceFunction = "cosineDistance")
         private final List<Float> embedding;
 
-        public GitHubFile() {
-            this(null, null, null, Collections.emptyList());
-        }
-
         public GitHubFile(
-            @JsonProperty("fileId") String id,
-            @JsonProperty("description") String description,
-            @JsonProperty("link") String link,
-            @JsonProperty("embedding") List<Float> embedding) {
+            String id,
+            String description,
+            String link,
+            List<Float> embedding) {
             this.id = id;
             this.description = description;
             this.link = link;
             this.embedding = embedding;
         }
 
+        public String getId() {
+            return id;
+        }
+        public String getDescription() {
+            return description;
+        }
+        public String getLink() {
+            return link;
+        }
+        public List<Float> getEmbedding() {
+            return embedding;
+        }
+
         static String encodeId(String realId) {
-            byte[] bytes = Base64.getUrlEncoder().encode(realId.getBytes(StandardCharsets.UTF_8));
-            return new String(bytes, StandardCharsets.UTF_8);
+            return VectorStoreWithAzureAISearch.GitHubFile.encodeId(realId);
         }
     }
 
     public static void main(String[] args) {
-        System.out.println("==============================================================");
-        System.out.println("========== Azure AI Search Vector Store Example ==============");
-        System.out.println("==============================================================");
+        System.out.println("===================================================================");
+        System.out.println("========== Volatile (In memory) Vector Store Example ==============");
+        System.out.println("===================================================================");
 
         OpenAIAsyncClient client;
 
@@ -104,33 +100,19 @@ public class AzureAISearchVectorStore {
             .withDimensions(EMBEDDING_DIMENSIONS)
             .build();
 
-        var searchClient = new SearchIndexClientBuilder()
-            .endpoint(AZURE_AI_SEARCH_ENDPOINT)
-            .credential(new AzureKeyCredential(AZURE_AISEARCH_KEY))
-            .clientOptions(clientOptions())
-            .buildAsyncClient();
-
-        dataStorageWithAzureAISearch(searchClient, embeddingGeneration);
+        inMemoryStoreAndSearch(embeddingGeneration);
     }
 
-    public static void dataStorageWithAzureAISearch(
-        SearchIndexAsyncClient searchClient,
+    public static void inMemoryStoreAndSearch(
         OpenAITextEmbeddingGenerationService embeddingGeneration) {
-
-        // Create a new Azure AI Search vector store
-        var azureAISearchVectorStore = com.microsoft.semantickernel.connectors.data.azureaisearch.AzureAISearchVectorStore
-            .builder()
-            .withSearchIndexAsyncClient(searchClient)
-            .withOptions(new AzureAISearchVectorStoreOptions())
-            .build();
+        // Create a new Volatile vector store
+        var volatileVectorStore = new VolatileVectorStore();
 
         String collectionName = "skgithubfiles";
-        var collection = (AzureAISearchVectorStoreRecordCollection<GitHubFile>) azureAISearchVectorStore
-            .getCollection(
-                collectionName,
-                AzureAISearchVectorStoreRecordCollectionOptions.<GitHubFile>builder()
-                    .withRecordClass(GitHubFile.class)
-                    .build());
+        var collection = volatileVectorStore.getCollection(collectionName,
+            VolatileVectorStoreRecordCollectionOptions.<GitHubFile>builder()
+                .withRecordClass(GitHubFile.class)
+                .build());
 
         // Create collection if it does not exist and store data
         collection
@@ -139,7 +121,7 @@ public class AzureAISearchVectorStore {
             .block();
 
         // Search for results
-        // Might need to wait for the data to be indexed
+        // Volatile store executes an exhaustive search, for approximate search use Azure AI Search, Redis or JDBC with PostgreSQL
         var results = search("How to get started", collection, embeddingGeneration).block();
 
         if (results == null || results.isEmpty()) {
@@ -148,21 +130,21 @@ public class AzureAISearchVectorStore {
         }
         var searchResult = results.get(0);
         System.out.printf("Search result with score: %f.%n Link: %s, Description: %s%n",
-            searchResult.getScore(), searchResult.getRecord().link,
-            searchResult.getRecord().description);
+                searchResult.getScore(), searchResult.getRecord().link,
+                searchResult.getRecord().description);
     }
 
     private static Mono<List<VectorSearchResult<GitHubFile>>> search(
-        String searchText,
-        AzureAISearchVectorStoreRecordCollection<GitHubFile> recordCollection,
-        OpenAITextEmbeddingGenerationService embeddingGeneration) {
-
+            String searchText,
+            VectorStoreRecordCollection<String, GitHubFile> recordCollection,
+            OpenAITextEmbeddingGenerationService embeddingGeneration) {
+        // Generate embeddings for the search text and search for the closest records
         return embeddingGeneration.generateEmbeddingsAsync(Collections.singletonList(searchText))
-            .flatMap(r -> recordCollection.searchAsync(r.get(0).getVector(), null));
+                .flatMap(r -> recordCollection.searchAsync(r.get(0).getVector(), null));
     }
 
     private static Mono<List<String>> storeData(
-        AzureAISearchVectorStoreRecordCollection<GitHubFile> recordCollection,
+        VectorStoreRecordCollection<String, GitHubFile> recordCollection,
         OpenAITextEmbeddingGenerationService embeddingGeneration,
         Map<String, String> data) {
 
@@ -170,6 +152,7 @@ public class AzureAISearchVectorStore {
             .flatMap(entry -> {
                 System.out.println("Save '" + entry.getKey() + "' to memory.");
 
+                // Generate embeddings for the data and store it
                 return embeddingGeneration
                     .generateEmbeddingsAsync(Collections.singletonList(entry.getValue()))
                     .flatMap(embeddings -> {
@@ -199,12 +182,5 @@ public class AzureAISearchVectorStore {
                 { "https://github.com/microsoft/semantic-kernel/blob/main/samples/apps/chat-summary-webapp-react/README.md",
                         "README: README associated with a sample chat summary react-based webapp" },
         }).collect(Collectors.toMap(element -> element[0], element -> element[1]));
-    }
-
-    private static ClientOptions clientOptions() {
-        return new ClientOptions()
-            .setTracingOptions(new TracingOptions())
-            .setMetricsOptions(new MetricsOptions())
-            .setApplicationId("Semantic-Kernel");
     }
 }
